Update Schema
Update sql schema script to automatically delete tables before attemping
to create them. This allows an easy clean slate for testing when test
information is added to the bottom of the script.

Add Persistent_Logins table to be used with the Spring Remember Me API
to remember recent logins of users.