Update Workflow JPA and add sample WorkflowDAO
Update Workflow JPA in an attempt to match the database schema. A new
class, WorkflowPK has been created to contain the primary key for the
workflow table.

Add test WorkflowDAO in order to see if the Workflow JPA is properly
mapped.

Update schema with some test information that can be used to see if the
JPA and DAO are working.

Rename WorkFlow to Workflow in all its occurances.

Note: system is not building