Add Javadoc, Update WorkflowService
Add Javadoc to Workflow, WorkflowPK, and AvailableTask. Rename TaskList
to AvailableTask to make it clear that this class represents tasks that
can be input by production users.

Update WorkflowDAO to take jobid as a parameter for finding a workflow.
Commented out methods left in if needed and to ensure system builds.