Add task_list preload entries to the data-mail.sql
SQL insert statements containing data were added to the sql script. This
will enable pre-population of the task_list table. All tasks were set to
billable and price to $0.00. This will be amended when the real billable
status and price are known. Note: the task_list table still needs to
have a department column added.