Implement AppRole and UserRole Repositories
Implement AppRole and UserRole repositories to replace the AppRole and
AppUser DAOs. Mark these DAOs as deprecated. Adjust
UserDetailsServiceImpl to use these new crud repositories instead.

Update UserRolePK class to be public to allow for searching for a
specific userrole by primary key using the new repository.

Update JobService to use the new JobRepository. JobService may need to
be updated to throw an exception in the future if a job isn't found
instead of returning null.