Implement input task execution flow
The flow from login>find job>create workflow> input task is working and
successfully inserts an input_task and workflow into the database. Changed
the mappings in WorkflowController, InputTaskController and their respective
html pages. Created a method in UserDetailsServiceImpl that retrieves an
AppUser by username. The username that of the logged in user.