Add workflow branch html to existing dev base
Added some html elements from the workflow branch manually. Added the menu
bar to all pages. This will be changed eventually because it makes no since
to give a user a link to a page they can't navigate to. At this point it
helps us demonstrate the User permission functionality of the project.
Added the New Workflow navigation functionality that exists on the workflow
branch. This allows a user to navigate to the create workflow page by pressing
the New Workflow button on the input task page. The workflow number is
updated in the frontend html for the button click.