Add validation
Added validation to cover if a user insert characters other than
numbers. It will delete characters as the user moves onto next input
box so that string will not be sent to server.If any input is left
blank, then focus will be set on the input. Also, cover  the case
when a default value of 0 is sent to the server, it will be converted
to null on the back-end.