Change PCR field to use a checkbox
Change is_pcr field on the inputTask page to be a checkbox instead of a
text input field. Since no value is sent to the controller when the
checkbox is left unchecked, the default value for the parameter in the
controller has been set to false.

Set parameter types to Integer from int in InputTaskController to
support optional parameters in the future.