Change TextField to Checkbox in TaskListManagement
Change textfield for isbillable value in AvailalableTask management to a
checkbox. This is done by enforcing a false value for isBillable when
the controller doesn't recieve a value from the webpage.

Remove redundant controller in AvailableTaskController. This controller
method provided the same functionality as saveTask. Some extra
validation may be need to be added in the future when an administrator
is updating tasks.